[1241] 1.Fix item system.stackable is max count in a slot.maxcount is total count one player can hold.0:no limit,1: only one.please change them to each other in your DB after checked. 2.Some small fix to HandleEffects.

+++ src/game/Player.h
@@ -525,7 +525,7 @@ class MANGOS_DLL_SPEC Player : public Unit
         bool SwapItem(uint8 dstBag,uint8 dstSlot,uint8 srcBag,uint8 srcSlot);
 
         Item* CreateNewItem (uint32 itemId, uint8 count = 1);
-        int GetItemCount(uint32 itemId, bool includebank = false);
+        uint16 GetItemCount(uint32 itemId, bool includebank = false);
         uint8 AddNewItem(uint32 itemId, uint32 count, bool addmaxpossible);
         uint8 AddItem(uint8 bagIndex, uint8 slot, Item *item, bool allowstack);
         uint8 AddItemToInventory(Item *item, bool addmaxpossible = false);
